streamlit/requirements.txt
# -------------------------------------
# Credit Risk Intelligence Platform
# Production Requirements (Optimized)
# -------------------------------------

# Web UI
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
pytz

# Visualization
plotly>=5.15.0

# Google Cloud BigQuery
google-cloud-bigquery>=3.11.0
google-cloud-core>=2.4.1
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
db-dtypes>=1.1.0

# Financial Data
yfinance>=0.2.18

# Arrow types required for BigQuery dataframe
pyarrow>=14.0.0
